+++ src/helpers.ts
@@ -151,9 +151,7 @@ export function updateBadge(treeView: vscode.TreeView<unknown>, logViewerProvide
     vscode.commands.executeCommand('setContext', 'magentoLogViewer.hasLogFiles', totalEntries > 0);
 
     // Update status bar item
-    const statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 100);
-    statusBarItem.text = `Magento Log-Entries: ${totalEntries}`;
-    statusBarItem.show();
+    LogViewerProvider.statusBarItem.text = `Magento Log-Entries: ${totalEntries}`;
   };
 
   logViewerProvider.onDidChangeTreeData(updateBadgeCount);
@@ -287,4 +285,34 @@ export function getIconForReport(filePath: string): vscode.ThemeIcon {
   } catch (error) {
     return new vscode.ThemeIcon('file');
   }
+}
+
+export function getReportItems(dir: string): LogItem[] {
+  if (!pathExists(dir)) {
+    return [];
+  }
+
+  const items: LogItem[] = [];
+  const files = fs.readdirSync(dir);
+
+  files.forEach(file => {
+    const filePath = path.join(dir, file);
+    if (fs.lstatSync(filePath).isDirectory()) {
+      const subItems = getReportItems(filePath);
+      if (subItems.length > 0) {
+        items.push(...subItems);
+      }
+    } else if (fs.lstatSync(filePath).isFile()) {
+      const title = parseReportTitle(filePath);
+      const reportFile = new LogItem(title, vscode.TreeItemCollapsibleState.None, {
+        command: 'magento-log-viewer.openFile',
+        title: 'Open Report File',
+        arguments: [filePath]
+      });
+      reportFile.iconPath = getIconForReport(filePath);
+      items.push(reportFile);
+    }
+  });
+
+  return items;
 }